fix: some remained merge conflicts

+++ include/graphics.h
@@ -1753,7 +1753,7 @@ extern const u32 gItemIcon_Pokevial8[];
 extern const u32 gItemIcon_Pokevial9[];
 extern const u32 gItemIcon_Pokevial[];
 
-extern const u32 gItemIconPalette_Pokevial[];
+extern const u16 gItemIconPalette_Pokevial[];
 // End Pokevial Branch
 
 extern const u32 gItemIcon_ReturnToFieldArrow[];
